#92: draft of the implementation
