+++ src/domain/creator/generatePortrait.ts
@@ -1,9 +1,9 @@
 
 import ImageData from '../image/data/ImageData';
-import generateImage from '../image/generateImage';
+import download from '../image/download';
 import { IMAGE_TYPE } from './data/constants';
 
 export default async function generatePortrait(pictureUrl: string): Promise<ImageData>
 {
-    return generateImage(pictureUrl, IMAGE_TYPE);
+    return download(pictureUrl, IMAGE_TYPE);
 }
